started ajax upload

+++ front/page/products.php
@@ -31,6 +31,9 @@ class Front_Page_Index extends Front_Page {
 			session_destroy();
 			header('Location: login');
 		}
+		if (!empty($_POST)) {
+			return '<pre>'.print_r($_POST,1).'</pre>';
+		}
 		if (isset($_POST['id'])) {
 			$item = front()->products()->getDetail($_POST['id']);
 			unlink(dirname(__FILE__).'/../../web'.$item['product_image']);
